+++ src/elastic-san/azext_elastic_san/aaz/latest/elastic_san/volume/_delete.py
@@ -29,9 +29,9 @@ class Delete(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2024-07-01-preview",
+        "version": "2025-09-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.elasticsan/elasticsans/{}/volumegroups/{}/volumes/{}", "2024-07-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.elasticsan/elasticsans/{}/volumegroups/{}/volumes/{}", "2025-09-01"],
         ]
     }
 
@@ -97,12 +97,6 @@ class Delete(AAZCommand):
                 max_length=63,
                 min_length=3,
             ),
-        )
-        _args_schema.delete_type = AAZStrArg(
-            options=["--delete-type"],
-            help="Optional. Specifies that the delete operation should be a permanent delete for the soft deleted volume. The value of deleteType can only be 'permanent'.",
-            is_preview=True,
-            enum={"permanent": "permanent"},
         )
         return cls._args_schema
 
@@ -200,10 +194,7 @@ class Delete(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "deleteType", self.ctx.args.delete_type,
-                ),
-                **self.serialize_query_param(
-                    "api-version", "2024-07-01-preview",
+                    "api-version", "2025-09-01",
                     required=True,
                 ),
             }